Added Workling-Transport.
You can now use Workling/Starling as transport
option. If you do so, the simple_publisher gem is
used to abstract a connection to a Starling queue
and send the arguments to that queue. The queue
name will be build according to Worklings naming
scheme from the service name and the name of the
service endpoint (its module name, to be precisely).