Introducing a client UUID to detect circular calls
In case no one (caller nor callee) is providing a
service implementation, you can get circular service
call cascades. In order to detect such a loop, I've
introduced a UUID, which is generated on Hoth.init!.
This UUID is included in the service call, so the
receiver can check if he has called himself.

TODOs:
* bert and workling transport does not make use of
  the caller UUID yet
* it would be much better, if you can detect such a
  circular call *BEFORE* making the call to self.
  Though I haven't figured out a nice way to do so
  yet.